+++ packages/react/src/index.ts
@@ -1,2 +1,12 @@
 export { Button, type ButtonProps } from "./button/button";
 export { Switch, type SwitchProps } from "./switch/switch";
+export {
+  TabsRoot,
+  TabsList,
+  TabsTab,
+  TabsPanel,
+  type TabsRootProps,
+  type TabsListProps,
+  type TabsTabProps,
+  type TabsPanelProps,
+} from "./tabs/tabs";